refactor: extract StorageRecoveryBanner.swift
Extracts the StorageRecoveryBanner View (12 lines) from
TranscriptRecoveryAndHistoryViews.swift into its own file.

Byte-identical move; SHA verified against origin/main.

Reduces TranscriptRecoveryAndHistoryViews.swift by 14 lines.

Tests: 667 (unchanged).
